+++ packages/react/src/modules/core/hooks/index.ts
@@ -1,2 +1,1 @@
 export { default as useOnBottomScroll } from "./useOnBottomScroll";
-export { default as useOutsideClick } from "./useOutsideClick";