+++ packages/react/src/modules/ms-teams/components/MsTeamsChannelCombobox/MsTeamsChannelInTeamCombobox.tsx
@@ -107,6 +107,10 @@ export const MsTeamsChannelInTeamCombobox: FunctionComponent<
           }
           closeOnSelect={false}
           layout="wrap"
+          modal={
+            // Modal comboboxes cause page layout to shift when body has padding. See KNO-7854.
+            false
+          }
         >
           <Combobox.Trigger />
           <Combobox.Content>