+++ packages/react/src/modules/slack/utils.ts
@@ -1,0 +1,8 @@
+import { SlackChannel } from "@knocklabs/client";
+
+export const sortSlackChannelsAlphabetically = (
+  channels: readonly SlackChannel[],
+) =>
+  [...channels].sort((channel1, channel2) =>
+    channel1.name.toLowerCase().localeCompare(channel2.name.toLowerCase()),
+  );