+++ packages/react-core/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.4.0
+
+### Minor Changes
+
+- 7904b65: Remove `slackSearchbarMultipleChannels` from translations strings
+
+## 0.3.4
+
+### Patch Changes
+
+- Updated dependencies [8ea25f4]
+  - @knocklabs/client@0.11.4
+
 ## 0.3.3
 
 ### Patch Changes